+++ Src/SOS.Import/Harvesters/TaxonListHarvester.cs
@@ -2,14 +2,13 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Net.Http;
 using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using SOS.Import.Harvesters.Interfaces;
-using SOS.Lib.Configuration.Shared;
 using SOS.Lib.Enums;
+using SOS.Lib.Managers.Interfaces;
 using SOS.Lib.Models.Verbatim.Shared;
 using SOS.Lib.Models.Shared;
 using SOS.Lib.Models.TaxonListService;
@@ -25,7 +24,7 @@ namespace SOS.Import.Harvesters
     {
         private readonly ITaxonListService _taxonListService;
         private readonly ITaxonListRepository _taxonListRepository;
-        private readonly SosApiConfiguration _sosApiConfiguration;
+        private readonly ICacheManager _cacheManager;
         private readonly ILogger<TaxonListHarvester> _logger;
 
         /// <summary>
@@ -33,43 +32,18 @@ namespace SOS.Import.Harvesters
         /// </summary>
         /// <param name="taxonListService"></param>
         /// <param name="taxonListRepository"></param>
-        /// <param name="sosApiConfiguration"></param>
+        /// <param name="cacheManager"></param>
         /// <param name="logger"></param>
         public TaxonListHarvester(
             ITaxonListService taxonListService,
             ITaxonListRepository taxonListRepository,
-            SosApiConfiguration sosApiConfiguration,
+            ICacheManager cacheManager,
             ILogger<TaxonListHarvester> logger)
         {
             _taxonListService = taxonListService ?? throw new ArgumentNullException(nameof(taxonListService));
             _taxonListRepository = taxonListRepository ?? throw new ArgumentNullException(nameof(taxonListRepository));
-            _sosApiConfiguration = sosApiConfiguration ?? throw new ArgumentNullException(nameof(sosApiConfiguration));
+            _cacheManager = cacheManager ?? throw new ArgumentNullException(nameof(cacheManager));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-        }
-
-        private async Task ClearTaxonListsCache()
-        {
-            try
-            {
-                var client = new HttpClient();
-                foreach (var observationsApiAddress in _sosApiConfiguration.ObservationsApiAddresses)
-                {
-                    var requestUri = $"{observationsApiAddress}Caches?cache={nameof(Cache.TaxonLists)}";
-                    var response = await client.DeleteAsync(requestUri);
-                    if (response.IsSuccessStatusCode)
-                    {
-                        _logger.LogInformation($"Taxon lists cache cleared ({observationsApiAddress})");
-                    }
-                    else
-                    {
-                        _logger.LogInformation($"Failed to clear taxon lists cache ({observationsApiAddress})");
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "Failed to clear taxon lists cache");
-            }
         }
 
         public async Task<HarvestInfo> HarvestTaxonListsAsync()
@@ -115,7 +89,8 @@ namespace SOS.Import.Harvesters
                 await _taxonListRepository.DeleteCollectionAsync();
                 await _taxonListRepository.AddCollectionAsync();
                 await _taxonListRepository.AddManyAsync(taxonLists);
-                await ClearTaxonListsCache();
+                // Clear observation api cache
+                await _cacheManager.ClearAsync(Cache.TaxonLists);
 
                 // Update harvest info
                 harvestInfo.End = DateTime.Now;