+++ Src/SOS.lib/Models/TaxonListService/TaxonListSetsById.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+
+namespace SOS.Lib.Models.TaxonListService
+{
+    /// <summary>
+    /// Only created to make it cacheable
+    /// </summary>
+    public class TaxonListSetsById : Dictionary<int, HashSet<int>>
+    {
+
+    }
+}