+++ Tests/Integration/SOS.Import.IntegrationTests/Harvesters/AreaHarvesterIntegrationTests.cs
@@ -8,7 +8,8 @@ using SOS.Lib.Configuration.Import;
 using SOS.Lib.Database;
 using SOS.Lib.Enums;
 using SOS.Lib.Helpers;
-using SOS.Lib.Repositories.Processed.Interfaces;
+using SOS.Lib.Managers;
+using SOS.Lib.Managers.Interfaces;
 using SOS.Lib.Repositories.Resource.Interfaces;
 using Xunit;
 
@@ -34,13 +35,16 @@ namespace SOS.Import.IntegrationTests.Harvesters
                     verbatimDbConfiguration.ReadBatchSize,
                     verbatimDbConfiguration.WriteBatchSize),
                 new Mock<ILogger<Lib.Repositories.Resource.AreaRepository>>().Object);
+            var cacheManager = new Mock<CacheManager>();
+            cacheManager.Setup(cm => cm.ClearAsync(Cache.Area)).ReturnsAsync(true);
 
             var areaHarvester = new AreaHarvester(
                 new Import.Repositories.Source.Artportalen.AreaRepository(artportalenDataService, new Mock<ILogger<Import.Repositories.Source.Artportalen.AreaRepository>>().Object),
                 areaVerbatimRepository,
                 new AreaHelper(new Mock<IAreaRepository>().Object),
                 new GeoRegionApiService(new GeoRegionApiConfiguration {ApiUrl = "https://georegionapi-dev.artdata.slu.se/"}), 
-                new AreaHarvestConfiguration {UseGeoRegionApiHarvest = true}, 
+                new AreaHarvestConfiguration {UseGeoRegionApiHarvest = true},
+                cacheManager.Object, 
                 new Mock<ILogger<AreaHarvester>>().Object);
 
             //-----------------------------------------------------------------------------------------------------------
@@ -72,12 +76,16 @@ namespace SOS.Import.IntegrationTests.Harvesters
                     verbatimDbConfiguration.WriteBatchSize),
                 new Mock<ILogger<Lib.Repositories.Resource.AreaRepository>>().Object);
 
+            var cacheManagerMock = new Mock<ICacheManager>();
+            cacheManagerMock.Setup(cm => cm.ClearAsync(Cache.Area)).ReturnsAsync(true);
+
             var areaHarvester = new AreaHarvester(
                 new Import.Repositories.Source.Artportalen.AreaRepository(artportalenDataService, new Mock<ILogger<Import.Repositories.Source.Artportalen.AreaRepository>>().Object),
                 areaVerbatimRepository,
                 new AreaHelper(new Mock<IAreaRepository>().Object),
                 new GeoRegionApiService(new GeoRegionApiConfiguration { ApiUrl = "https://georegionapi-dev.artdata.slu.se/" }),
                 new AreaHarvestConfiguration { UseGeoRegionApiHarvest = true },
+                cacheManagerMock.Object,
                 new Mock<ILogger<AreaHarvester>>().Object);
 
             //-----------------------------------------------------------------------------------------------------------